# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Steward System",
        "description": "Documentação",
        "contact": {
            "email": "anaimayer3@gmail.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1.0"
    },
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "d55f416d61f094eb6038b566bc6e4596",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "fulano@exemplo.com.br",
                                    "password": "S4af3P4assword"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwODAvYXBpL2xvZ2luIiwiaWF0IjoxNzE0MDU4NDI1LCJleHAiOjE3MTQ5MjI0MjUsIm5iZiI6MTcxNDA1ODQyNSwianRpIjoiVGxZUzhQYXFpNDdKQnhpdiIsInN1YiI6IjMiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.2xKUA7vbL80YHU9lKv4pKbXSTUC7lk-GVgfb-nLDxhY",
                                        "token_type": "Bearer",
                                        "expires_in": 14400
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Usuário ou senha incorreta"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable content"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "7f261a1e14b78ceb6cadbe969582a855",
                "responses": {
                    "204": {
                        "description": "No content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Token de acesso destruído"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Token could not be parsed from the request."
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh",
                "operationId": "79723d033278f41c747dc3af98965ca4",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwODAvYXBpL2xvZ2luIiwiaWF0IjoxNzE0MDU4NDI1LCJleHAiOjE3MTQ5MjI0MjUsIm5iZiI6MTcxNDA1ODQyNSwianRpIjoiVGxZUzhQYXFpNDdKQnhpdiIsInN1YiI6IjMiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.2xKUA7vbL80YHU9lKv4pKbXSTUC7lk-GVgfb-nLDxhY"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Token could not be parsed from the request."
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "The token has been blacklisted"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/get-me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get me",
                "operationId": "7e47871f94badc1376c1749e74f45c29",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "usuario": {
                                            "type": "json"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "usuario": {
                                            "uuid": "7a02d1a0-1bd7-4d55-a32d-05e6ad3bee10",
                                            "email": "exemplo@mail.com.br",
                                            "email_verified_at": null,
                                            "change_password": true,
                                            "created_at": "2024-04-25T11:43:40.000000Z",
                                            "updated_at": "2024-04-25T11:43:40.000000Z",
                                            "admin": false,
                                            "perfil": {
                                                "nome": "Usuário Exemplo",
                                                "cpf": "61141289075",
                                                "cpf_responsavel": false,
                                                "telefone": "48998673412",
                                                "altura": "1.75",
                                                "avatar_url": "https://t3.ftcdn.net/jpg/05/53/79/60/360_F_553796090_XHrE6R9jwmBJUMo9HKl41hyHJ5gqt9oz.jpg"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Token inválido/Não encontrado"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}